+++ src/settings/ReadingRoomAccess/constant.js
@@ -1,0 +1,6 @@
+// eslint-disable-next-line import/prefer-default-export
+export const readingRoomAccessColumns = {
+  NAME: 'name',
+  ISPUBLIC: 'isPublic',
+  SERVICEPOINTS: 'servicePoints',
+};